Move 'derived_variables' to 'ops'
