+++ example.danra.yaml
@@ -82,11 +82,11 @@ inputs:
           time: time
           lat: lat
           lon: lon
-        function: mllam_data_prep.derived_variables.calculate_toa_radiation
+        function: mllam_data_prep.ops.derived_variables.calculate_toa_radiation
       hour_of_day:
         kwargs:
           time: time
-        function: mllam_data_prep.derived_variables.calculate_hour_of_day
+        function: mllam_data_prep.ops.derived_variables.calculate_hour_of_day
     dim_mapping:
       time:
         method: rename
